Move tsdb-labels into its own tool
